+++ CHANGES.md
@@ -26,6 +26,7 @@ Changes:
 - Retrieve lower court information in `wyo` #1569
 - Retrieve lower court information in `dc` #1569
 - Retrieve lower court information in `cal` scrapers #1569
+- Retrieve lower court information in `delaware` scrapers #1569
 
 Fixes:
 - Fix connappct #1580